put the props.onAddEditDeleteComplete behind the if + selectedDateSlot (fix)
put the props.onAddEditDeleteComplete behind the if. Cuz you can do a whole CRUD operation at the page 'Vacation', 'Holiday' and 'NurseShift' itself OR the 'Planning' calendar page where the props.onAddComplete(), props.onEditComplete, and props.onDeleteComplete() counts.

I have no other idea how to work around it.

If I didn't put that in, it also would work but my selectedModal wouldn't close if I submited the Add/Update/Delete form modal. And I rather want to close it too after the submit.

And

Pass props selectedDateSlot issue fix

+++ PlanningsTool/planningstool.ui/src/components/holidays/DeleteHoliday.js
@@ -55,7 +55,9 @@ function DeleteHoliday(props) {
                 toast.error('Feestdag is verwijderd');
                 handleClose();
                 props.onUpdate();
-                props.onDeleteComplete();
+                if (props.onDeleteComplete) {
+                    props.onDeleteComplete();
+                }
             })
             .catch((error) => {
                 toast.warning(`${error}`);
