+++ PlanningsTool/planningstool.ui/src/components/nurseShifts/DeleteNurseShift.js
@@ -113,7 +113,9 @@ function DeleteNurseShift(props) {
       .then(() => {
         toast.error('Nurse shift is verwijderd');
         props.onUpdate();
-        props.onDeleteComplete();
+        if (props.onDeleteComplete) {
+          props.onDeleteComplete();
+        }
         handleClose();
       })
       .catch((error) => {
