+++ PlanningsTool/planningstool.ui/src/components/vacations/DeleteVacation.js
@@ -105,7 +105,9 @@ function DeleteVacation(props) {
                 toast.error('Verlof is verwijderd');
                 handleClose();
                 props.onUpdate();
-                props.onDeleteComplete();
+                if(props.onDeleteComplete){
+                    props.onDeleteComplete();
+                }
             })
             .catch((error) => {
                 toast.warning(`${error}`);
